Convert over to using closure manager
