--- Source/GenerateSharp/PackageManager.Core/ClosureBuilder.cs
@@ -1,22 +1,0 @@
-﻿// <copyright file="ClosureBuilder.cs" company="Soup">
-// Copyright (c) Soup. All rights reserved.
-// </copyright
-
-namespace Soup.Build.PackageManager
-{
-	using System;
-	using System.Collections.Generic;
-	using System.Linq;
-	using System.Net.Http;
-	using System.Threading.Tasks;
-	using Opal;
-	using Opal.System;
-	using Soup.Build.Utilities;
-
-	/// <summary>
-	/// The closure builder
-	/// </summary>
-	public class ClosureBuilder
-	{
-	}
-}
