+++ Source/GenerateSharp/PackageManager.Core/PackageManager.cs
@@ -5,7 +5,6 @@
 namespace Soup.Build.PackageManager
 {
 	using System;
-	using System.Collections.Generic;
 	using System.Linq;
 	using System.Net.Http;
 	using System.Threading.Tasks;
@@ -19,31 +18,21 @@ namespace Soup.Build.PackageManager
 	public class PackageManager
 	{
 		private const string StagingFolderName = ".staging/";
-		private const string RootClosureName = "Root";
-		private const string BuiltInLanguageCSharp = "C#";
-		private const string BuiltInLanguageCpp = "C++";
-		private const string BuiltInLanguagePackageCSharp = "Soup.CSharp";
-		private const string BuiltInLanguagePackageCpp = "Soup.Cpp";
-		private const string BuiltInLanguageSafeNameCSharp = "CSharp";
-		private const string BuiltInLanguageSafeNameCpp = "Cpp";
 
 		private Uri _apiEndpoint;
 
 		private HttpClient _httpClient;
 
-		private SemanticVersion _builtInLanguageVersionCSharp;
-		private SemanticVersion _builtInLanguageVersionCpp;
+		private IClosureManager _closureManager;
 
 		public PackageManager(
 			Uri apiEndpoint,
 			HttpClient httpClient,
-			SemanticVersion builtInLanguageVersionCSharp,
-			SemanticVersion builtInLanguageVersionCpp)
+			IClosureManager closureManager)
 		{
 			_apiEndpoint = apiEndpoint;
 			_httpClient = httpClient;
-			_builtInLanguageVersionCSharp = builtInLanguageVersionCSharp;
-			_builtInLanguageVersionCpp = builtInLanguageVersionCpp;
+			_closureManager = closureManager;
 		}
 
 		/// <summary>
@@ -63,18 +52,7 @@ namespace Soup.Build.PackageManager
 
 			try
 			{
-				// Place the lock in the local directory
-				var packageLockPath =
-					workingDirectory +
-					BuildConstants.PackageLockFileName;
-
-				await CheckRestorePackageClosureAsync(
-					workingDirectory,
-					packageLockPath,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingPath);
+				await _closureManager.GenerateAndRestoreRecursiveLocksAsync(workingDirectory, packageStore, stagingPath);
 
 				// Cleanup the working directory
 				Log.Diag("Deleting staging directory");
@@ -150,19 +128,6 @@ namespace Soup.Build.PackageManager
 
 				if (targetPackageReference.Version == null)
 					throw new InvalidOperationException("Target package version was null");
-
-				var closure = new Dictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>>();
-				var buildClosures = new Dictionary<string, IDictionary<string, IDictionary<string, PackageReference>>>();
-				await CheckRecursiveEnsurePackageDownloadedAsync(
-					recipe.Language.Name,
-					targetPackageReference.Name,
-					targetPackageReference.Version,
-					false,
-					packageStore,
-					lockStore,
-					stagingDirectory,
-					closure,
-					buildClosures);
 
 				// Cleanup the working directory
 				Log.Info("Deleting staging directory");
@@ -361,721 +326,6 @@ namespace Soup.Build.PackageManager
 		}
 
 		/// <summary>
-		/// Ensure a package version is downloaded
-		/// </summary>
-		private async Task CheckRecursiveEnsurePackageDownloadedAsync(
-			string languageName,
-			string packageName,
-			SemanticVersion packageVersion,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory,
-			IDictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>> closure,
-			IDictionary<string, IDictionary<string, IDictionary<string, PackageReference>>> buildClosures)
-		{
-			if (closure.ContainsKey(languageName) && closure[languageName].ContainsKey(packageName))
-			{
-				Log.Diag("Recipe already processed.");
-			}
-			else
-			{
-				await EnsurePackageDownloadedAsync(
-					languageName,
-					packageName,
-					packageVersion,
-					packageStore,
-					stagingDirectory);
-
-				var languageRootFolder = packageStore + new Path(languageName);
-				var packageRootFolder = languageRootFolder + new Path(packageName);
-				var packageVersionFolder = packageRootFolder + new Path(packageVersion.ToString()) + new Path("/");
-
-				var recipePath =
-					packageVersionFolder +
-					BuildConstants.RecipeFileName;
-				var (isSuccess, recipe) = await RecipeExtensions.TryLoadRecipeFromFileAsync(recipePath);
-				if (!isSuccess)
-				{
-					throw new InvalidOperationException($"Could not load the recipe file: {recipePath}");
-				}
-
-				// Create the unique build closure
-				var buildClosure = await CreateBuildClosureAsync(recipe, packageVersionFolder);
-
-				var buildClosureName = string.Empty;
-				var match = buildClosures.FirstOrDefault(value => AreEqual(value.Value, buildClosure));
-				if (match.Key != null)
-				{
-					buildClosureName = match.Key;
-				}
-				else
-				{
-					buildClosureName = $"Build{buildClosures.Count}";
-					buildClosures.Add(buildClosureName, buildClosure);
-				}
-
-				// Add the new package to the closure
-				if (!closure.ContainsKey(languageName))
-					closure.Add(languageName, new Dictionary<string, (PackageReference Package, string BuildClosure)>());
-				closure[languageName].Add(packageName, (new PackageReference(null, packageName, packageVersion), buildClosureName));
-
-				// Install recursive dependencies
-				await RestoreRecursiveDependenciesAsync(
-					packageVersionFolder,
-					recipe,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory,
-					closure,
-					buildClosures);
-			}
-		}
-
-		/// <summary>
-		/// Ensure a package version is downloaded
-		/// </summary>
-		private async Task EnsurePackageDownloadedAsync(
-			string languageName,
-			string packageName,
-			SemanticVersion packageVersion,
-			Path packageStore,
-			Path stagingDirectory)
-		{
-			Log.HighPriority($"Install Package: {languageName} {packageName}@{packageVersion}");
-
-			var languageRootFolder = packageStore + new Path(languageName);
-			var packageRootFolder = languageRootFolder + new Path(packageName);
-			var packageVersionFolder = packageRootFolder + new Path(packageVersion.ToString()) + new Path("/");
-
-			// Check if the package version already exists
-			if ((packageName == BuiltInLanguagePackageCpp && packageVersion == _builtInLanguageVersionCpp) ||
-				(packageName == BuiltInLanguagePackageCSharp && packageVersion == _builtInLanguageVersionCSharp))
-			{
-				Log.HighPriority("Skip built in language version");
-			}
-			else if (LifetimeManager.Get<IFileSystem>().Exists(packageVersionFolder))
-			{
-				Log.HighPriority("Found local version");
-			}
-			else
-			{
-				// Download the archive
-				Log.HighPriority("Downloading package");
-				var archivePath = stagingDirectory + new Path(packageName + ".zip");
-
-				var client = new Api.Client.PackageVersionsClient(_httpClient)
-				{
-					BaseUrl = _apiEndpoint.ToString(),
-				};
-
-				try
-				{
-					var result = await client.DownloadPackageVersionAsync(languageName, packageName, packageVersion.ToString());
-
-					// Write the contents to disk, scope cleanup
-					using (var archiveWriteFile = LifetimeManager.Get<IFileSystem>().OpenWrite(archivePath, true))
-					{
-						await result.Stream.CopyToAsync(archiveWriteFile.GetOutStream());
-					}
-				}
-				catch (Api.Client.ApiException ex)
-				{
-					if (ex.StatusCode == 404)
-					{
-						Log.HighPriority("Package Version Missing");
-						throw new HandledException();
-					}
-					else
-					{
-						throw;
-					}
-				}
-
-				// Create the package folder to extract to
-				var stagingVersionFolder = stagingDirectory + new Path($"{languageName}_{packageName}_{packageVersion}/");
-				LifetimeManager.Get<IFileSystem>().CreateDirectory2(stagingVersionFolder);
-
-				// Unpack the contents of the archive
-				LifetimeManager.Get<IZipManager>().ExtractToDirectory(archivePath, stagingVersionFolder);
-
-				// Delete the archive file
-				LifetimeManager.Get<IFileSystem>().DeleteFile(archivePath);
-
-				// Ensure the package root folder exists
-				if (!LifetimeManager.Get<IFileSystem>().Exists(packageRootFolder))
-				{
-					// Create the folder
-					LifetimeManager.Get<IFileSystem>().CreateDirectory2(packageRootFolder);
-				}
-
-				// Move the extracted contents into the version folder
-				LifetimeManager.Get<IFileSystem>().Rename(stagingVersionFolder, packageVersionFolder);
-			}
-		}
-
-		/// <summary>
-		/// Restore package lock
-		/// </summary>
-		private async Task RestorePackageLockAsync(
-			Path packageStore,
-			Path stagingDirectory,
-			PackageLock packageLock)
-		{
-			foreach (var closure in packageLock.GetClosures().Values)
-			{
-				Log.Info($"Restore Packages for Closure {closure.Key}");
-				foreach (var languageProjects in closure.Value.Value.AsTable().Values)
-				{
-					var languageName = GetLanguageName(languageProjects.Key);
-					Log.Info($"Restore Packages for Language {languageName}");
-					foreach (var project in languageProjects.Value.Value.AsArray().Values)
-					{
-						var projectTable = project.Value.AsTable();
-						var projectName = projectTable.Values[PackageLock.Property_Name].Value.AsString().Value;
-						var projectVersion = projectTable.Values[PackageLock.Property_Version].Value.AsString().Value;
-						if (SemanticVersion.TryParse(projectVersion, out var version))
-						{
-							await EnsurePackageDownloadedAsync(
-								languageName,
-								projectName,
-								version,
-								packageStore,
-								stagingDirectory);
-						}
-						else
-						{
-							Log.Info($"Skip Package: {projectName} -> {projectVersion}");
-						}
-					}
-				}
-			}
-		}
-
-		/// <summary>
-		/// Restore packages
-		/// </summary>
-		public async Task CheckRestorePackageClosureAsync(
-			Path workingDirectory,
-			Path packageLockPath,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingPath)
-		{
-			bool wasRestored = false;
-			if (!forceRestore)
-			{
-				var loadPackageLock = await PackageLockExtensions.TryLoadFromFileAsync(packageLockPath);
-				if (loadPackageLock.IsSuccess)
-				{
-					Log.Info("Restore from package lock");
-					await RestorePackageLockAsync(
-						packageStore,
-						stagingPath,
-						loadPackageLock.Result);
-					wasRestored = true;
-				}
-			}
-
-			if (!wasRestored)
-			{
-				Log.Info("Discovering full closure");
-				var closure = new Dictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>>();
-				var buildClosures = new Dictionary<string, IDictionary<string, IDictionary<string, PackageReference>>>();
-				await DiscoverRecursiveDependenciesAsync(
-					workingDirectory,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingPath,
-					closure,
-					buildClosures);
-
-				// Build up the package lock file
-				var packageLock = new PackageLock();
-				packageLock.SetVersion(3);
-				foreach (var languageClosure in closure.OrderBy(value => value.Key))
-				{
-					var languageSafeName = GetLanguageSafeName(languageClosure.Key);
-					foreach (var (key, (package, buildClosure)) in languageClosure.Value.OrderBy(value => value.Key))
-					{
-						var value = string.Empty;
-						if (package.IsLocal)
-						{
-							value = package.Path.GetRelativeTo(workingDirectory).ToString();
-						}
-						else
-						{
-							if (package.Version == null)
-								throw new InvalidOperationException("Package lock closure must have version");
-							value = package.Version.ToString();
-						}
-
-						Log.Diag($"{RootClosureName}:{languageClosure.Key} {key} -> {value}");
-						packageLock.AddProject(
-							RootClosureName,
-							languageSafeName,
-							key,
-							value,
-							buildClosure);
-					}
-				}
-
-				foreach (var buildClosure in buildClosures.OrderBy(value => value.Key))
-				{
-					packageLock.EnsureClosure(buildClosure.Key);
-					foreach (var languageClosure in buildClosure.Value.OrderBy(value => value.Key))
-					{
-						var languageSafeName = GetLanguageSafeName(languageClosure.Key);
-						foreach (var (key, package) in languageClosure.Value)
-						{
-							var value = string.Empty;
-							if (package.IsLocal)
-							{
-								value = package.Path.GetRelativeTo(workingDirectory).ToString();
-							}
-							else
-							{
-								if (package.Version == null)
-									throw new InvalidOperationException("Package lock closure must have version");
-								value = package.Version.ToString();
-							}
-
-							Log.Diag($"{buildClosure.Key}:{languageClosure.Key} {key} -> {value}");
-							packageLock.AddProject(
-								buildClosure.Key,
-								languageSafeName,
-								key,
-								value,
-								null);
-						}
-					}
-				}
-
-				// Save the updated package lock
-				await PackageLockExtensions.SaveToFileAsync(packageLockPath, packageLock);
-			}
-		}
-
-		/// <summary>
-		/// Recursively restore all dependencies
-		/// </summary>
-		private async Task DiscoverRecursiveDependenciesAsync(
-			Path recipeDirectory,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory,
-			IDictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>> closure,
-			IDictionary<string, IDictionary<string, IDictionary<string, PackageReference>>> buildClosures)
-		{
-			var recipePath =
-				recipeDirectory +
-				BuildConstants.RecipeFileName;
-			var (isSuccess, recipe) = await RecipeExtensions.TryLoadRecipeFromFileAsync(recipePath);
-			if (!isSuccess)
-			{
-				throw new InvalidOperationException($"Could not load the recipe file: {recipePath}");
-			}
-
-			if (closure.TryGetValue(recipe.Language.Name, out var languageClosure) && languageClosure.ContainsKey(recipe.Name))
-			{
-				Log.Diag("Recipe already processed.");
-			}
-			else
-			{
-				// Create the unique build closure
-				var buildClosure = await CreateBuildClosureAsync(recipe, recipeDirectory);
-
-				var buildClosureName = string.Empty;
-				var match = buildClosures.FirstOrDefault(value => AreEqual(value.Value, buildClosure));
-				if (match.Key != null)
-				{
-					buildClosureName = match.Key;
-				}
-				else
-				{
-					buildClosureName = $"Build{buildClosures.Count}";
-					buildClosures.Add(buildClosureName, buildClosure);
-				}
-
-				// Add the project to the closure
-				if (!closure.ContainsKey(recipe.Language.Name))
-					closure.Add(recipe.Language.Name, new Dictionary<string, (PackageReference Package, string BuildClosure)>());
-				closure[recipe.Language.Name].Add(recipe.Name, (new PackageReference(recipeDirectory), buildClosureName));
-
-				await RestoreRecursiveDependenciesAsync(
-					recipeDirectory,
-					recipe,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory,
-					closure,
-					buildClosures);
-			}
-		}
-
-		private static async Task<IDictionary<string, IDictionary<string, PackageReference>>> CreateBuildClosureAsync(
-			Recipe recipe,
-			Path recipeDirectory)
-		{
-			var buildClosure = new Dictionary<string, IDictionary<string, PackageReference>>();
-			var implicitLanguage = BuiltInLanguageCSharp;
-
-			// Add the language build extension
-			var recipeLanguagePackage = GetLanguagePackage(recipe.Language.Name);
-			buildClosure.Add(implicitLanguage, new Dictionary<string, PackageReference>());
-			buildClosure[implicitLanguage].Add(
-				recipeLanguagePackage,
-				FillDefaultVersion(new PackageReference(implicitLanguage, recipeLanguagePackage, recipe.Language.Version)));
-
-			// Discover any dependency build references
-			if (recipe.HasBuildDependencies)
-			{
-				foreach (var dependency in recipe.BuildDependencies)
-				{
-					PackageReference dependencyPackage;
-					string dependencyName;
-					string dependencyLanguage;
-					if (dependency.IsLocal)
-					{
-						// Load the recipe to check for the language and name of the package
-						var dependencyPath = recipeDirectory + dependency.Path;
-						var dependencyRecipePath =
-							dependencyPath +
-							BuildConstants.RecipeFileName;
-						var (isDependencySuccess, dependencyRecipe) =
-							await RecipeExtensions.TryLoadRecipeFromFileAsync(dependencyRecipePath);
-						if (!isDependencySuccess)
-						{
-							throw new InvalidOperationException("Could not load dependency recipe file.");
-						}
-
-						dependencyPackage = dependency;
-						dependencyName = dependencyRecipe.Name;
-						dependencyLanguage = dependencyRecipe.Language.Name;
-					}
-					else
-					{
-						dependencyPackage = FillDefaultVersion(dependency);
-						dependencyName = dependencyPackage.Name;
-						dependencyLanguage = dependencyPackage.Language != null ? dependencyPackage.Language : implicitLanguage;
-					}
-
-					if (!buildClosure.ContainsKey(dependencyLanguage))
-						buildClosure.Add(dependencyLanguage, new Dictionary<string, PackageReference>());
-
-					buildClosure[dependencyLanguage].Add(dependencyName, dependencyPackage);
-				}
-			}
-
-			return buildClosure;
-		}
-
-		private static string GetLanguageSafeName(string language)
-		{
-			switch (language)
-			{
-				case BuiltInLanguageCSharp:
-					return BuiltInLanguageSafeNameCSharp;
-				case BuiltInLanguageCpp:
-					return BuiltInLanguageSafeNameCpp;
-				default:
-					throw new InvalidOperationException("Unknown language");
-			}
-		}
-
-		private static string GetLanguagePackage(string language)
-		{
-			switch (language)
-			{
-				case BuiltInLanguageCSharp:
-					return BuiltInLanguagePackageCSharp;
-				case BuiltInLanguageCpp:
-					return BuiltInLanguagePackageCpp;
-				default:
-					throw new InvalidOperationException("Unknown language");
-			}
-		}
-
-		private static string GetLanguageName(string languageSafeName)
-		{
-			switch (languageSafeName)
-			{
-				case BuiltInLanguageSafeNameCSharp:
-					return BuiltInLanguageCSharp;
-				case BuiltInLanguageSafeNameCpp:
-					return BuiltInLanguageCpp;
-				default:
-					throw new InvalidOperationException("Unknown language safe name");
-			}
-		}
-
-		private static PackageReference FillDefaultVersion(PackageReference package)
-		{
-			if (package.Version == null)
-				throw new ArgumentException("Package version was null");
-
-			// TODO: Discover the latest available version
-			// For now auto assume missing values are zero
-			if (package.Version.Minor is null)
-			{
-				return new PackageReference(
-					package.Language,
-					package.Name,
-					new SemanticVersion(package.Version.Major, 0, 0));
-			}
-			else if (package.Version.Patch is null)
-			{
-				return new PackageReference(
-					package.Language,
-					package.Name,
-					new SemanticVersion(package.Version.Major, package.Version.Minor, 0));
-			}
-			else
-			{
-				return package;
-			}
-		}
-
-		/// <summary>
-		/// Recursively restore all dependencies, assume that the closure has been updated correctly for current recipe
-		/// </summary>
-		private async Task RestoreRecursiveDependenciesAsync(
-			Path recipeDirectory,
-			Recipe recipe,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory,
-			IDictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>> closure,
-			IDictionary<string, IDictionary<string, IDictionary<string, PackageReference>>> buildClosures)
-		{
-			// Restore the language build extension
-			var recipeLanguagePackage = GetLanguagePackage(recipe.Language.Name);
-			var recipeLanguagePackageReference = FillDefaultVersion(new PackageReference(null, recipeLanguagePackage, recipe.Language.Version));
-
-			if ((recipeLanguagePackageReference.Name == BuiltInLanguagePackageCpp && recipeLanguagePackageReference.Version == _builtInLanguageVersionCpp) ||
-				(recipeLanguagePackageReference.Name == BuiltInLanguagePackageCSharp && recipeLanguagePackageReference.Version == _builtInLanguageVersionCSharp))
-			{
-				Log.Info("Skip Built In Language Build Dependency");
-			}
-			else
-			{
-				Log.Info("Restore Language Build Dependency");
-				await RestoreBuildDependencyAsync(
-					recipeDirectory,
-					recipeLanguagePackageReference,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory);
-			}
-
-			// Restore the explicit dependencies
-			foreach (var dependencyType in recipe.GetDependencyTypes())
-			{
-				// Build dependencies do not inherit the parent language
-				// Instead, they default to C#
-				bool isBuildDependency = dependencyType == Recipe.Property_Build;
-				if (isBuildDependency)
-				{
-					await RestoreBuildDependenciesAsync(
-						recipeDirectory,
-						recipe,
-						forceRestore,
-						packageStore,
-						lockStore,
-						stagingDirectory);
-				}
-				else
-				{
-					await RestoreRuntimeDependenciesAsync(
-						recipeDirectory,
-						recipe,
-						dependencyType,
-						forceRestore,
-						packageStore,
-						lockStore,
-						stagingDirectory,
-						closure,
-						buildClosures);
-				}
-			}
-		}
-
-		/// <summary>
-		/// Recursively restore all runtime dependencies
-		/// </summary>
-		private async Task RestoreRuntimeDependenciesAsync(
-			Path recipeDirectory,
-			Recipe recipe,
-			string dependencyType,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory,
-			IDictionary<string, IDictionary<string, (PackageReference Package, string BuildClosure)>> closure,
-			IDictionary<string, IDictionary<string, IDictionary<string, PackageReference>>> buildClosures)
-		{
-			// Same language as parent is implied
-			var implicitLanguage = recipe.Language.Name;
-
-			foreach (var dependency in recipe.GetNamedDependencies(dependencyType))
-			{
-				// If local then check children for external package references
-				// Otherwise install the external package reference and its dependencies
-				if (dependency.IsLocal)
-				{
-					var dependencyPath = recipeDirectory + dependency.Path;
-					await DiscoverRecursiveDependenciesAsync(
-						dependencyPath,
-						forceRestore,
-						packageStore,
-						lockStore,
-						stagingDirectory,
-						closure,
-						buildClosures);
-				}
-				else
-				{
-					if (dependency.Version == null)
-						throw new ArgumentException("Local package version was null");
-
-					var language = dependency.Language != null ? dependency.Language : implicitLanguage;
-					await CheckRecursiveEnsurePackageDownloadedAsync(
-						language,
-						dependency.Name,
-						dependency.Version,
-						forceRestore,
-						packageStore,
-						lockStore,
-						stagingDirectory,
-						closure,
-						buildClosures);
-				}
-			}
-		}
-
-		/// <summary>
-		/// Recursively restore all build dependencies
-		/// </summary>
-		private async Task RestoreBuildDependenciesAsync(
-			Path recipeDirectory,
-			Recipe recipe,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory)
-		{
-			foreach (var dependency in recipe.GetNamedDependencies(Recipe.Property_Build))
-			{
-				await RestoreBuildDependencyAsync(
-					recipeDirectory,
-					dependency,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory);
-			}
-		}
-
-		/// <summary>
-		/// Recursively restore a build dependency
-		/// </summary>
-		private async Task RestoreBuildDependencyAsync(
-			Path workingDirectory,
-			PackageReference dependencyReference,
-			bool forceRestore,
-			Path packageStore,
-			Path lockStore,
-			Path stagingDirectory)
-		{
-			// Build dependencies do not inherit the parent language
-			// Instead, they default to C#
-			var implicitLanguage = BuiltInLanguageCSharp;
-
-			// If local then check children for external package references
-			// Otherwise install the external package reference and its dependencies
-			if (dependencyReference.IsLocal)
-			{
-				var dependencyPath = workingDirectory + dependencyReference.Path;
-
-				// Place the lock in the local directory
-				var packageLockPath =
-					dependencyPath +
-					BuildConstants.PackageLockFileName;
-
-				await CheckRestorePackageClosureAsync(
-					dependencyPath,
-					packageLockPath,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory);
-			}
-			else
-			{
-				if (dependencyReference.Version == null)
-					throw new ArgumentException("Local package version was null");
-
-				var language = dependencyReference.Language != null ? dependencyReference.Language : implicitLanguage;
-
-				await EnsurePackageDownloadedAsync(
-					language,
-					dependencyReference.Name,
-					dependencyReference.Version,
-					packageStore,
-					stagingDirectory);
-
-				var packageLanguageNameVersionPath = new Path(language) +
-					new Path(dependencyReference.Name) +
-					new Path(dependencyReference.Version.ToString()) + new Path("/");
-				var packageContentDirectory = packageStore + packageLanguageNameVersionPath;
-
-				// Place the lock in the lock store
-				var packageLockDirectory =
-					lockStore +
-					packageLanguageNameVersionPath;
-				var packageLockPath =
-					packageLockDirectory +
-					BuildConstants.PackageLockFileName;
-
-				EnsureDirectoryExists(packageLockDirectory);
-
-				await CheckRestorePackageClosureAsync(
-					packageContentDirectory,
-					packageLockPath,
-					forceRestore,
-					packageStore,
-					lockStore,
-					stagingDirectory);
-			}
-		}
-
-		private static bool AreEqual(
-			IDictionary<string, IDictionary<string, PackageReference>> lhs,
-			IDictionary<string, IDictionary<string, PackageReference>> rhs)
-		{
-			return lhs.Keys.Count == rhs.Keys.Count &&
-				lhs.Keys.All(value => rhs.Keys.Contains(value)) &&
-				lhs.All(value => AreEqual(value.Value, rhs[value.Key]));
-		}
-
-		private static bool AreEqual(
-			IDictionary<string, PackageReference> lhs,
-			IDictionary<string, PackageReference> rhs)
-		{
-
-			return lhs.Keys.Count == rhs.Keys.Count &&
-				lhs.Keys.All(value => rhs.Keys.Contains(value)) &&
-				lhs.All(value => value.Value == rhs[value.Key]);
-		}
-
-		/// <summary>
 		/// Ensure the staging directory exists
 		/// </summary>
 		static Path EnsureStagingDirectoryExists(Path packageStore)
@@ -1091,19 +341,6 @@ namespace Soup.Build.PackageManager
 			LifetimeManager.Get<IFileSystem>().CreateDirectory2(stagingDirectory);
 
 			return stagingDirectory;
-		}
-
-		/// <summary>
-		/// Ensure the staging directory exists
-		/// </summary>
-		static void EnsureDirectoryExists(Path directory)
-		{
-			if (!LifetimeManager.Get<IFileSystem>().Exists(directory))
-			{
-				// Create the folder
-				Log.Diag($"Create Directory: {directory}");
-				LifetimeManager.Get<IFileSystem>().CreateDirectory2(directory);
-			}
 		}
 	}
 }
