+++ Source/GenerateSharp/PackageManager.UnitTests/Soup.Build.PackageManager.UnitTests.csproj
@@ -4,6 +4,9 @@
 		<IsPackable>false</IsPackable>
 		<Platforms>AnyCPU;x64</Platforms>
 	</PropertyGroup>
+	<PropertyGroup>
+		<LangVersion>preview</LangVersion>
+	</PropertyGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.0" />
 		<PackageReference Include="Moq" Version="4.18.2" />
